명령어 정리

+++ nodejs/chatbot/src/telegram/command_chatgpt.ts
@@ -4,6 +4,7 @@ import path from "path";
 import { getOpenAI } from "../lib/openai";
 import telegramBot from '../lib/telegramBot';
 import { isEmpty } from "lodash";
+import UserRole from "../lib/userRole";
 
 const logger = BaseLogger.createLogger(path.basename(__filename, path.extname(__filename)));
 
@@ -11,7 +12,11 @@ export namespace CommandChatGPT {
 
     export const Name = "/chatgpt";
 
-    export const ForceReplyText = `${CommandChatGPT.Name} 프롬프트를 입력해주세요.`;
+    export const ForceReplyText = `${Name} 프롬프트를 입력해주세요.`;
+
+    export enum CallbackTypes {
+        
+    }
 
     const defaultModel = "gpt-3.5-turbo";
 
@@ -29,7 +34,16 @@ export namespace CommandChatGPT {
             return;
         }
 
-        let prompt;
+        if(!await UserRole.isPermission('chatgpt', msg.chat.id, msg.from)){
+            telegramBot.getBot().sendMessage(msg.chat.id, `권한이 없습니다.\n>> user: ${msg.from.username}(${msg.from.id})`,{
+                reply_to_message_id: msg.message_id
+            }).catch(err=>{
+                logger.error(`telegramBot.getBot().deleteMessage:(): error. ${err}`);
+            });
+            return;
+        }
+
+        let prompt:string;
         if(msg.reply_to_message?.text === ForceReplyText){
             telegramBot.getBot().deleteMessage(msg.chat.id, msg.reply_to_message.message_id).catch(err=>{
                 logger.error(`telegramBot.getBot().deleteMessage:(): error. ${err}`);
@@ -62,15 +76,6 @@ export namespace CommandChatGPT {
             return;
         }
 
-        if(msg.from.id != 58087716){
-            telegramBot.getBot().sendMessage(msg.chat.id, "안알랴줌.", {
-                reply_to_message_id: msg.message_id
-            }).catch(err=>{
-                logger.error(`telegramBot.getBot().sendMessage:(): error. ${err}`);
-            })
-            return;
-        }
-        
         const rMsg = await telegramBot.getBot().sendMessage(msg.chat.id, "GPT에게 물어보는 중...", {
             reply_to_message_id: msg.message_id
         });
@@ -105,8 +110,10 @@ export namespace CommandChatGPT {
         }
     }
 
-    export function onCallbackQuery(_bot:TelegramBot, _query:TelegramBot.CallbackQuery, _args:string[]){
-        
+    export function onCallbackQuery(bot:TelegramBot, query:TelegramBot.CallbackQuery, _args:string[]){
+        bot.answerCallbackQuery(query.id).catch(err=>{
+            logger.error(`bot.answerCallbackQuery(): error. ${err}`);
+        });
     }
 }
